Fix bug 688932 mswinpr2 doesn't clear temporary files. Unwanted temporary file was being recreated during page output. We now delete the temporary file when the device is closed.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7477 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevwpr2.c
@@ -392,12 +392,13 @@ win_pr2_open(gx_device * dev)
 
     /* gdev_prn_open opens a temporary file which we don't want */
     /* so we specify the name now so we can delete it later */
+    wdev->fname[0] = '\0';
     pfile = gp_open_scratch_file(gp_scratch_file_name_prefix,
 				 wdev->fname, "wb");
     fclose(pfile);
     code = gdev_prn_open(dev);
-    /* delete unwanted temporary file */
-    unlink(wdev->fname);
+    if ((code < 0) && wdev->fname[0])
+        unlink(wdev->fname);
 
     if (!wdev->nocancel) {
 	/* inform user of progress with dialog box and allow cancel */
@@ -444,6 +445,11 @@ win_pr2_close(gx_device * dev)
     }
 
     code = gdev_prn_close(dev);
+
+    /* delete unwanted temporary file */
+    if (wdev->fname[0])
+	unlink(wdev->fname);
+
     return code;
 }
